# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Burgos National High School Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current session user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated

  # ── DASHBOARD ─────────────────────────────────────────────────────────────
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      parameters:
        - name: pinned
          in: query
          schema:
            type: ["boolean", "null"]
        - name: emergency
          in: query
          schema:
            type: ["boolean", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementListResponse"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    get:
      operationId: getAnnouncement
      tags: [announcements]
      summary: Get announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── EVENTS ────────────────────────────────────────────────────────────────
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: upcoming
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── DOWNLOADS ─────────────────────────────────────────────────────────────
  /downloads:
    get:
      operationId: listDownloads
      tags: [downloads]
      summary: List downloadable files
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Downloads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Download"
    post:
      operationId: createDownload
      tags: [downloads]
      summary: Create download entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DownloadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Download"
  /downloads/{id}:
    patch:
      operationId: updateDownload
      tags: [downloads]
      summary: Update download entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DownloadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Download"
    delete:
      operationId: deleteDownload
      tags: [downloads]
      summary: Delete download entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── GALLERY ───────────────────────────────────────────────────────────────
  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery items
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: media_type
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"
    post:
      operationId: createGalleryItem
      tags: [gallery]
      summary: Add gallery item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
  /gallery/{id}:
    delete:
      operationId: deleteGalleryItem
      tags: [gallery]
      summary: Delete gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── SSLG ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, username, name, role]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, teacher, staff]

    # ── DASHBOARD ─────────────────────────────────────────────────────────
...
    Announcement:
      type: object
      required: [id, title, content, isPinned, isEmergency, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean
        isEmergency:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
    AnnouncementInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean
          default: false
        isEmergency:
          type: boolean
          default: false
        expiresAt:
          type: string
    AnnouncementUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean
        isEmergency:
          type: boolean
        expiresAt:
          type: string
    AnnouncementListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Announcement"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── EVENTS ────────────────────────────────────────────────────────────
    Event:
      type: object
      required: [id, title, startDate, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        createdAt:
          type: string
    EventInput:
      type: object
      required: [title, startDate]
      properties:
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        location:
          type: string
        category:
          type: string
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        location:
          type: string
        category:
          type: string

    # ── DOWNLOADS ─────────────────────────────────────────────────────────
    Download:
      type: object
      required: [id, title, category, fileName, fileUrl, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        fileName:
          type: string
        fileUrl:
          type: string
        createdAt:
          type: string
    DownloadInput:
      type: object
      required: [title, category, fileName, fileUrl]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        fileName:
          type: string
        fileUrl:
          type: string
    DownloadUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        fileName:
          type: string
        fileUrl:
          type: string

    # ── GALLERY ───────────────────────────────────────────────────────────
    GalleryItem:
      type: object
      required: [id, title, mediaType, url, category, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        mediaType:
          type: string
          enum: [photo, video]
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        category:
          type: string
        createdAt:
          type: string
    GalleryInput:
      type: object
      required: [title, mediaType, url, category]
      properties:
        title:
          type: string
        description:
          type: string
        mediaType:
          type: string
          enum: [photo, video]
        url:
          type: string
        thumbnailUrl:
          type: string
        category:
          type: string

    # ── SSLG ──────────────────────────────────────────────────────────────
...