# lib/api-zod/src/generated/types/announcementUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Burgos National High School Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AnnouncementUpdate {
  title?: string;
  content?: string;
  isPinned?: boolean;
  isEmergency?: boolean;
  expiresAt?: string;
}